move enums back to godotsteam
